+++ cpskin/core/viewlets/related_contacts.py
@@ -7,8 +7,11 @@ from collective.geo.mapwidget import utils
 from cpskin.core.utils import format_phone
 from plone import api
 from plone.app.layout.viewlets import common
+from plone.dexterity.utils import safe_unicode
 from plone.dexterity.utils import safe_utf8
-from plone.outputfilters.filters.resolveuid_and_caption import ResolveUIDAndCaptionFilter  # noqa
+from plone.outputfilters.filters.resolveuid_and_caption import (
+    ResolveUIDAndCaptionFilter,
+)  # noqa
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile
 from pygeoif.geometry import as_shape
 
@@ -17,24 +20,39 @@ import logging
 import Missing
 
 
-logger = logging.getLogger('cpskin.core related contacts viewlet')
+logger = logging.getLogger("cpskin.core related contacts viewlet")
 
 
 class RelatedContactsViewlet(common.ViewletBase):
 
-    index = ViewPageTemplateFile('related_contacts.pt')
-    address_fields = ('street', 'number', 'zip_code', 'city',
-                      'additional_address_details', 'region', 'country')
-    coordinates_fields = ('phone', 'cell_phone', 'fax', 'email', 'im_handle',
-                          'website', 'schedule')
-    ignore_fields = ('title', )
+    index = ViewPageTemplateFile("related_contacts.pt")
+    address_fields = (
+        "street",
+        "number",
+        "zip_code",
+        "city",
+        "additional_address_details",
+        "region",
+        "country",
+    )
+    coordinates_fields = (
+        "phone",
+        "cell_phone",
+        "fax",
+        "email",
+        "im_handle",
+        "website",
+        "schedule",
+    )
+    ignore_fields = ("title",)
 
-    def __init__(self, context, request, view, manager=None, field=[], selected=[]):  # noqa
-        super(RelatedContactsViewlet, self).__init__(
-            context, request, view, manager)
+    def __init__(
+        self, context, request, view, manager=None, field=[], selected=[]
+    ):  # noqa
+        super(RelatedContactsViewlet, self).__init__(context, request, view, manager)
         self.field = field
         self.selected = selected
-        self.pc = api.portal.get_tool('portal_catalog')
+        self.pc = api.portal.get_tool("portal_catalog")
 
     def available(self):
         contacts = getattr(self.context, self.field, None)
@@ -62,8 +80,10 @@ class RelatedContactsViewlet(common.ViewletBase):
         return contacts
 
     def get_title(self, contact):
-        if self.in_fields('title'):
-            return u'<span class="related-contact-title">{0}</span>'.format(contact.title)
+        if self.in_fields("title"):
+            return u'<span class="related-contact-title">{0}</span>'.format(
+                contact.title
+            )
         else:
             return False
 
@@ -71,47 +91,48 @@ class RelatedContactsViewlet(common.ViewletBase):
         return field in self.selected_fields
 
     def get_field(self, contact, field_name):
-        if field_name is 'id':
-            return getattr(contact, field_name, '')
+        if field_name is "id":
+            return getattr(contact, field_name, "")
         if field_name not in self.selected_fields:
-            return ''
+            return ""
         if field_name in self.address_fields:
             contactable = IContactable(contact)
             details = contactable.get_contact_details()
-            return details['address'].get(field_name)
+            return details["address"].get(field_name)
         # field = getattr(contact, field_name, '')
         # find way to check if field is richetext or image or simple field
-        if getattr(getattr(contact, field_name, ''), 'raw', None):
-            if getattr(contact, field_name, ''):
+        if getattr(getattr(contact, field_name, ""), "raw", None):
+            if getattr(contact, field_name, ""):
                 text = getattr(contact, field_name).raw
-                text = text.replace('http://resolveuid/', 'resolveuid/')
+                text = text.replace("http://resolveuid/", "resolveuid/")
                 parser = ResolveUIDAndCaptionFilter(contact)
                 transform_text = parser(text)
-                return transform_text if transform_text else ''
-        if field_name in ['logo', 'photo']:
-            if getattr(contact, field_name, ''):
-                img = contact.unrestrictedTraverse('@@images')
+                return transform_text if transform_text else ""
+        if field_name in ["logo", "photo"]:
+            if getattr(contact, field_name, ""):
+                img = contact.unrestrictedTraverse("@@images")
                 logo = img.scale(field_name)
-                return logo.tag() if logo.tag() else ''
-        if field_name == 'schedule':
+                return logo.tag() if logo.tag() else ""
+        if field_name == "schedule":
             from plone.directives import dexterity
+
             display = dexterity.DisplayForm(contact, self.request)
             display.update()
-            if display.w.get('IScheduledContent.schedule', None):
-                return display.w.get('IScheduledContent.schedule').render()
+            if display.w.get("IScheduledContent.schedule", None):
+                return display.w.get("IScheduledContent.schedule").render()
             else:
-                return ''
-        if field_name in ['phone', 'cell_phone', 'fax']:
-            phones = getattr(contact, field_name, '')
+                return ""
+        if field_name in ["phone", "cell_phone", "fax"]:
+            phones = getattr(contact, field_name, "")
             if not phones:
                 return False
             if not isinstance(phones, list):
                 phones = [getattr(contact, field_name)]
             return [format_phone(phone) for phone in phones]
-        if field_name in ['position']:
+        if field_name in ["position"]:
             positions = [pos.title.strip() for pos in contact.get_held_positions()]
-            return ', '.join(positions)
-        return getattr(contact, field_name, '')
+            return ", ".join(positions)
+        return getattr(contact, field_name, "")
 
     def has_address(self):
         i = 0
@@ -123,26 +144,29 @@ class RelatedContactsViewlet(common.ViewletBase):
     def fields_without_address(self):
         fields = []
         for selected_field in self.selected_fields:
-            if selected_field not in self.address_fields and \
-                    selected_field not in self.ignore_fields and \
-                    selected_field not in self.coordinates_fields:
+            if (
+                selected_field not in self.address_fields
+                and selected_field not in self.ignore_fields
+                and selected_field not in self.coordinates_fields
+            ):
                 fields.append(selected_field)
         return fields
 
     def get_website(self, contact):
-        website = self.get_field(contact, 'website')
-        if website.startswith('http'):
+        website = self.get_field(contact, "website")
+        if website.startswith("http"):
             url = website
-            website_name = website.replace('http://', '')
-        elif website.startswith('https'):
+            website_name = website.replace("http://", "")
+        elif website.startswith("https"):
             url = website
-            website_name = website.replace('https://', '')
+            website_name = website.replace("https://", "")
         else:
-            url = 'http://{0}'.format(website)
+            url = "http://{0}".format(website)
             website_name = website
-        html = ''
+        html = ""
         html += '<a class="website" href="{0}" target="_blank">{1}</a>'.format(
-            url, website_name)
+            url, website_name
+        )
         return html
 
     def see_map_link(self, contact):
@@ -153,65 +177,49 @@ class RelatedContactsViewlet(common.ViewletBase):
             brain = self.pc.unrestrictedSearchResults(UID=obj.UID())[0]
             if brain.zgeo_geometry == Missing.Value:
                 return False
-            if not getattr(self.context, 'see_map', True):
+            if not getattr(self.context, "see_map", True):
                 return False
             return True
         return False
 
 
 class AboveRelatedContactsViewlet(RelatedContactsViewlet):
-
     def __init__(self, context, request, view, manager=None):
-        field = 'aboveContentContact'
-        selected = 'aboveVisbileFields'
+        field = "aboveContentContact"
+        selected = "aboveVisbileFields"
         super(AboveRelatedContactsViewlet, self).__init__(
-            context,
-            request,
-            view,
-            manager,
-            field,
-            selected
+            context, request, view, manager, field, selected
         )
 
 
 class BelowRelatedContactsViewlet(RelatedContactsViewlet):
-
     def __init__(self, context, request, view, manager=None):
-        field = 'belowContentContact'
-        selected = 'belowVisbileFields'
+        field = "belowContentContact"
+        selected = "belowVisbileFields"
         super(BelowRelatedContactsViewlet, self).__init__(
-            context,
-            request,
-            view,
-            manager,
-            field,
-            selected
+            context, request, view, manager, field, selected
         )
 
     def get_title(self, contact):
-        if self.in_fields('title'):
+        if self.in_fields("title"):
             return u'<a href="{0}" target="_blank"><h4>{1}</h4></a>'.format(
-                contact.absolute_url(),
-                contact.title)
+                contact.absolute_url(), safe_unicode(contact.title)
+            )
         else:
             return False
 
 
 class RelatedContactsMapViewlet(RelatedContactsViewlet):
-    index = ViewPageTemplateFile('related_contacts_map.pt')
+    index = ViewPageTemplateFile("related_contacts_map.pt")
 
     def __init__(self, context, request, view, manager=None):
-        self.fields = ['aboveContentContact', 'belowContentContact']
+        self.fields = ["aboveContentContact", "belowContentContact"]
         super(RelatedContactsMapViewlet, self).__init__(
-            context,
-            request,
-            view,
-            manager,
-            self.fields
+            context, request, view, manager, self.fields
         )
 
     def available(self):
-        see_map = getattr(self.context, 'see_map', False)
+        see_map = getattr(self.context, "see_map", False)
         empty_content = True
         for field in self.fields:
             if len(getattr(self.context, field, [])) > 0:
@@ -225,68 +233,74 @@ class RelatedContactsMapViewlet(RelatedContactsViewlet):
     def data_geojson(self):
         style = {}
         global_style = utils.get_feature_styles(self.context)
-        style['fill'] = global_style['polygoncolor']
-        style['stroke'] = global_style['linecolor']
-        style['width'] = global_style['linewidth']
-        if global_style.get('marker_image', None):
-            img = get_marker_image(self.context, global_style['marker_image'])
-            style['image'] = img
+        style["fill"] = global_style["polygoncolor"]
+        style["stroke"] = global_style["linecolor"]
+        style["width"] = global_style["linewidth"]
+        if global_style.get("marker_image", None):
+            img = get_marker_image(self.context, global_style["marker_image"])
+            style["image"] = img
         else:
-            style['image'] = None
+            style["image"] = None
         json_result = []
-        self.pc = api.portal.get_tool('portal_catalog')
+        self.pc = api.portal.get_tool("portal_catalog")
         for contact in self.get_contacts():
             brain = self.pc.unrestrictedSearchResults(UID=contact.UID())[0]
             if contact.use_parent_address:
-                brain = self.pc.unrestrictedSearchResults(
-                    UID=contact.aq_parent.UID())[0]
+                brain = self.pc.unrestrictedSearchResults(UID=contact.aq_parent.UID())[
+                    0
+                ]
             if brain.zgeo_geometry == Missing.Value:
                 continue
             if brain.collective_geo_styles == Missing.Value:
                 continue
-            if brain.collective_geo_styles.get('use_custom_styles', False) and \
-                   brain.collective_geo_styles.get('marker_image', None):
-                img = get_marker_image(self.context, brain.collective_geo_styles['marker_image'])
-                style['image'] = img
-            geom = {'type': brain.zgeo_geometry['type'],
-                    'coordinates': brain.zgeo_geometry['coordinates']}
-            if geom['coordinates']:
-                if geom['type']:
-                    classes = geom['type'].lower() + ' '
-                else:
-                    classes = ''
-                address = get_address(contact)
-                number = ''
-                if address.get('number', None):
-                    number = ', {0}'.format(address['number'])
-                formated_address = '{0}{1}<br />{2} {3}'.format(
-                    safe_utf8(address.get('street') or ''),
-                    number,
-                    address.get('zip_code') or '',
-                    safe_utf8(address.get('city') or '')
+            if brain.collective_geo_styles.get(
+                "use_custom_styles", False
+            ) and brain.collective_geo_styles.get("marker_image", None):
+                img = get_marker_image(
+                    self.context, brain.collective_geo_styles["marker_image"]
                 )
-                img = ''
+                style["image"] = img
+            geom = {
+                "type": brain.zgeo_geometry["type"],
+                "coordinates": brain.zgeo_geometry["coordinates"],
+            }
+            if geom["coordinates"]:
+                if geom["type"]:
+                    classes = geom["type"].lower() + " "
+                else:
+                    classes = ""
+                address = get_address(contact)
+                number = ""
+                if address.get("number", None):
+                    number = ", {0}".format(address["number"])
+                formated_address = "{0}{1}<br />{2} {3}".format(
+                    safe_utf8(address.get("street") or ""),
+                    number,
+                    address.get("zip_code") or "",
+                    safe_utf8(address.get("city") or ""),
+                )
+                img = ""
                 if self.context.see_logo_in_popup:
-                    acc = getattr(contact, 'logo', None)
+                    acc = getattr(contact, "logo", None)
                     if acc and acc.filename:
-                        img = '{0}/@@images/logo/thumb'.format(
-                            contact.absolute_url()
-                        )
-                classes += brain.getPath().split('/')[-2].replace('.', '-')
+                        img = "{0}/@@images/logo/thumb".format(contact.absolute_url())
+                classes += brain.getPath().split("/")[-2].replace(".", "-")
                 json_result.append(
                     geojson.Feature(
-                        id=contact.id.replace('.', '-'),
+                        id=contact.id.replace(".", "-"),
                         geometry=as_shape(geom),
                         style=style,
                         properties={
-                            'title': brain.Title,
-                            'description': brain.Description,
-                            'style': style,
-                            'url': brain.getURL(),
-                            'classes': classes,
-                            'image': img,
-                            'address': formated_address
-                        }))
+                            "title": brain.Title,
+                            "description": brain.Description,
+                            "style": style,
+                            "url": brain.getURL(),
+                            "classes": classes,
+                            "image": img,
+                            "address": formated_address,
+                        },
+                    )
+                )
         feature_collection = geojson.FeatureCollection(json_result)
-        feature_collection.update({'title': self.context.title})
+        feature_collection.update({"title": self.context.title})
         return geojson.dumps(feature_collection)